Serialisation generators for Mary/Allegra
Resolves CAD-1845

* Extract EraIndepGenerators from the ShelleyEra serialisation generators so that we can re-use EraIndep as a base for ShelleyMAEra serialisation generators
* Specialise Shelley Era generators to ShelleyEra (rather than just the `ShelleyTest` constraint) - otherwise we'd have Overlapping Instances between Shelley/MA
* ShelleyMAEra generators are fixed to MaryEra C_Crypto and AllegraEra C_Crypto

Note: To import Shelley + ShelleyMA arb instances, you need

```
import Test.Shelley.Spec.Ledger.Serialisation.EraIndepGenerators () -- shelley-spec-ledger-test
import Test.Shelley.Spec.Ledger.Serialisation.Generators ()         -- shelley-spec-ledger-test
import Test.Cardano.Ledger.ShelleyMA.Serialisation.Generators ()    -- cardano-ledger-shelley-ma-test
```
